+++ package.json
@@ -1,30 +1,36 @@
 {
-  "homepage": "https://rudrpsingh.tech",
   "name": "rudr-pratap-singh",
-  "version": "0.1.0",
+  "version": "0.1.1",
+  "homepage": "https://rudrpsingh.tech",
   "private": true,
   "dependencies": {
-    "@emailjs/browser": "^3.11.0",
-    "@emotion/react": "^11.11.0",
+    "@emailjs/browser": "^4.1.0",
+    "@emotion/react": "^11.11.3",
     "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.11.16",
-    "@mui/lab": "^5.0.0-alpha.130",
-    "@mui/material": "^5.13.1",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.1.1",
+    "@mui/icons-material": "^5.15.10",
+    "@mui/lab": "^5.0.0-alpha.165",
+    "@mui/material": "^5.15.10",
+    "@react-three/drei": "^9.99.0",
+    "@react-three/fiber": "^8.15.16",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "@vercel/analytics": "^1.3.1",
-    "@vercel/speed-insights": "^1.0.12",
-    "emailjs": "^4.0.2",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0",
-    "react-icons": "^4.3.1",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "^5.0.1",
-    "react-scroll": "^1.8.7",
-    "styled-components": "^5.3.5",
-    "typewriter-effect": "^2.19.0",
-    "web-vitals": "^2.1.4"
+    "emailjs": "^4.0.3",
+    "framer-motion": "^11.0.5",
+    "maath": "^0.10.7",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^5.0.1",
+    "react-router-dom": "^6.22.1",
+    "react-scripts": "5.0.1",
+    "react-scroll": "^1.9.0",
+    "react-tilt": "^1.0.2",
+    "react-vertical-timeline-component": "^3.6.0",
+    "styled-components": "^6.1.8",
+    "three": "^0.161.0",
+    "typewriter-effect": "^2.21.0",
+    "web-vitals": "^2.1.4",
+    "gh-pages": "^3.2.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -50,8 +56,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "gh-pages": "^6.1.1"
   }
 }
