--- src/components/About/index.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const About = () => {
-  return (
-    <div>About</div>
-  )
-}
-
-export default About;
