+++ src/components/HeroBgAnimation/HeroBgAnimationStyle.js
