+++ src/components/HeroBgAnimation/index.js
